Move blocking traits to the crate root
