+++ src/pwm.rs
@@ -34,7 +34,7 @@
 /// # impl U32Ext for u32 { fn khz(self) -> KiloHertz { KiloHertz(self) } }
 /// # enum Channel { _1, _2 }
 /// # struct Pwm1;
-/// # impl hal::blocking::pwm::Pwm for Pwm1 {
+/// # impl hal::pwm::Pwm for Pwm1 {
 /// #     type Error = Infallible;
 /// #     type Channel = Channel;
 /// #     type Time = KiloHertz;
