+++ src/rng.rs
